Feature: Extreme changes + working handshakes
Next up is receiving game packets, dispatching them to listeners, and handling reliability internally

+++ src/main/kotlin/raknet/packet/DataPacket.kt
@@ -4,7 +4,7 @@ import io.netty.buffer.ByteBuf
 import io.netty.buffer.ByteBufAllocator
 import raknet.codec.encode
 
-abstract class DataPacket(val id: Short) : Packet {
+abstract class DataPacket(val id: Int) : Packet {
 
     override fun encode(): ByteBuf {
         val buffer = ByteBufAllocator.DEFAULT.ioBuffer()
@@ -13,14 +13,14 @@ abstract class DataPacket(val id: Short) : Packet {
     }
 
     /**
-     * We don't use decode at the moment as we use Packet.from() as a way to decode the packet
+     * We don't use decode at the moment as we use XPacket.from() as a way to decode the packet
      * It may be worth a look at using decode in the future
      */
     override fun decode(buffer: ByteBuf) = Unit
 
     abstract fun encodeOrder(): Array<Any>
 
-    open fun encodeHeader(buffer: ByteBuf): ByteBuf = buffer.writeByte(id.toInt())
+    open fun encodeHeader(buffer: ByteBuf): ByteBuf = buffer.writeByte(id)
 
     fun prepare(): ByteBuf {
         val encoded = encode()
